Adding the Contributor Covenant Code of Conduct
See #6. I think that this PR should add it here, and then that we should add it to the other main documentation repositories. I did not see the need to use a different CoC - I think this one should do fine. Note that I also added myself to the moderation team, because I do not like having only one person on it, and I am happy to do this work as needed.

@komaeda what do you think?